requirements.txt
Flask>=3.1.2
PyYAML>=6.0.2
requests>=2.32.5
gunicorn>=23.0.0